pyproject.toml
[project]
name = "tg-service-v2"
version = "0.1.0"
description = "TAD Telegram Service v2 — Kafka-compatible replacement for tg-messenger-service"
requires-python = ">=3.12"
dependencies = [
    # Kafka
    "aiokafka>=0.10.0",
    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.0",
    # Redis
    "redis[hiredis]>=5.0.0",
    # Telegram
    "aiogram>=3.19.0",
    "telethon>=1.39.0",
    "pysocks>=1.7.1",
    "aiohttp-socks>=0.10.1",
    # Config & Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Security
    "cryptography>=42.0.0",
    # Logging & Monitoring
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "sentry-sdk>=2.5.0",
    # Retry & Resilience
    "tenacity>=8.3.0",
    # Fast JSON (3-10x faster than stdlib, native UUID/datetime)
    "orjson>=3.10.0",
    # CLI
    "typer[all]>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "*.egg-info",
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "ARG002", # unused method argument
    "ARG001", # unused function argument
    "RUF002",
    "RUF003",
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
